chore: add temporary experimental formula for testing mac binaries
This commit adds a new formula that we can use to temporarily
test changes to the homebrew update/release process until we
have it all working.  That way we won't break the main one that
folks might still currently be using.

The end goal is to have a formula that pulls the mac/linux
binaries directly from the github release and doesn't require
re-building them either in homebrew or on the user's machine.