chore: restore uploading artifacts at end of test run
When an auto-generated PR comes in from the other repo, after
the test phase runs, the publish phase wants to pull artifacts
off of the github run from the test phase.  In a previous commit
I removed those and that caused the build to fail.  This
attempts to restore them.